feat(projen.project.nx-monorepo): move and refactor nx-monorepo to subproject
