+++ packages/projen/project/nx-monorepo/src/index.ts
@@ -1,0 +1,2 @@
+export * from './nx-monorepo'
+export * from './nx-monorepo-project-options'
