+++ packages/projen/project/nx-monorepo/src/nx-monorepo-project-options.ts
@@ -1,6 +1,6 @@
 // ~~ Generated by projen. To modify, edit .projenrc.ts and run "npx projen".
-import { github, GitOptions, IgnoreFileOptions, javascript, LoggerOptions, Project, ProjenrcJsonOptions, release, RenovatebotOptions, SampleReadmeProps, typescript } from 'projen';
-import { MonorepoUpgradeDepsOptions, WorkspaceConfig } from '@aws-prototyping-sdk/nx-monorepo';
+import type { github, GitOptions, IgnoreFileOptions, javascript, LoggerOptions, Project, ProjenrcJsonOptions, release, RenovatebotOptions, SampleReadmeProps, typescript } from 'projen';
+import type { MonorepoUpgradeDepsOptions, WorkspaceConfig } from '@aws-prototyping-sdk/nx-monorepo';
 
 /**
  * NxMonorepoProjectOptions
