+++ projenrc/project.ts
@@ -4,25 +4,12 @@ import {
 	ReleasePlease,
 	ReleaseType,
 } from '@arroyodev-llc/projen.component.release-please'
-import { ToolVersions } from '@arroyodev-llc/projen.component.tool-versions'
 import { UnBuild } from '@arroyodev-llc/projen.component.unbuild'
-import {
-	Vitest,
-	VitestConfigType,
-} from '@arroyodev-llc/projen.component.vitest'
 import { Vue } from '@arroyodev-llc/projen.component.vue'
-import { findComponent } from '@arroyodev-llc/utils.projen'
-import nx_monorepo, { NodePackageUtils } from '@aws-prototyping-sdk/nx-monorepo'
-import {
-	cdk,
-	type github,
-	javascript,
-	JsonFile,
-	type Project,
-	typescript,
-} from 'projen'
-import { NodePackage, TypeScriptModuleResolution } from 'projen/lib/javascript'
-import type { NxMonorepoProjectOptions } from './nx-monorepo-project-options'
+import { type MonorepoProject } from '@arroyodev-llc/projen.project.nx-monorepo'
+import { ProjectName } from '@arroyodev-llc/utils.projen'
+import { NodePackageUtils } from '@aws-prototyping-sdk/nx-monorepo'
+import { cdk, javascript, typescript } from 'projen'
 import type { ProjenProjectOptions } from './projen-project-options'
 import type { TypeScriptProjectOptions } from './typescript-project-options'
 
@@ -48,226 +35,6 @@ const projectDefaults = {
 	repositoryUrl: 'https://github.com/arroyodev-llc/components',
 	projenrcTs: true,
 } satisfies ProjenProjectOptions
-
-export class MonorepoProject extends nx_monorepo.NxMonorepoProject {
-	public readonly esmBundledTsconfigExtends: javascript.TypescriptConfigExtends
-
-	constructor(options: NxMonorepoProjectOptions) {
-		const { workspaceDeps, tsconfigBase, tsconfig, ...rest } = options
-		super({
-			...projectDefaults,
-			releaseToNpm: false,
-			projenDevDependency: true,
-			tsconfig,
-			...rest,
-		})
-		const pnpmWorkspace = new PnpmWorkspace(this)
-		pnpmWorkspace.addWorkspaceDeps(...(workspaceDeps ?? []))
-		new LintConfig(this)
-		this.gitignore.exclude('.idea', '.idea/**')
-		this.addDevDeps('tsx')
-		this.defaultTask!.reset('tsx .projenrc.ts')
-		new JsonFile(this, '.ncurc.json', {
-			readonly: true,
-			marker: false,
-			allowComments: false,
-			obj: {
-				reject: [],
-			},
-		})
-		new Vitest(this, {
-			configType: VitestConfigType.WORKSPACE,
-			settings: {
-				test: {
-					threads: true,
-				},
-			},
-		})
-		this.esmBundledTsconfigExtends = this.buildEsmBundledTsConfig()
-		this.applyNpmConfig(
-			findComponent(this, javascript.NpmConfig) ??
-				new javascript.NpmConfig(this)
-		)
-			.applyGithub(this.github!)
-			.applyPackage(this.package)
-		this.tsconfigDev!.addExtends(this.tsconfig!)
-		// readonly access token (safe to be public)
-		this.nx.useNxCloud(
-			'NTc0NTE5MGItNjY3Ni00YmQzLTg0YTUtNWFkMzc5ZWZiY2Y4fHJlYWQtb25seQ=='
-		)
-		this.nx.autoInferProjectTargets = true
-		new ToolVersions(this, {
-			tools: {
-				nodejs: ['18.15.0', 'lts'],
-				pnpm: ['8.5.1'],
-			},
-		})
-	}
-
-	protected applyCleanTask(): this {
-		this.addDevDeps('rimraf')
-		const cleanAllTask =
-			this.tasks.tryFind('clean') ?? this.tasks.addTask('clean')
-		cleanAllTask.exec('pnpm -r --parallel run clean')
-		return this.applyRecursive((project) => {
-			const nodePackage = findComponent(project, NodePackage)
-			if (nodePackage) {
-				nodePackage.setScript(
-					'clean',
-					NodePackageUtils.command.projen(nodePackage.packageManager, 'clean')
-				)
-			}
-			const cleanTask =
-				project.tasks.tryFind('clean') ?? project.tasks.addTask('clean')
-			cleanTask.exec(
-				NodePackageUtils.command.exec(
-					this.package.packageManager,
-					'rimraf dist'
-				)
-			)
-			cleanTask.exec(
-				NodePackageUtils.command.exec(this.package.packageManager, 'rimraf lib')
-			)
-		})
-	}
-
-	protected applyGithub(gh: github.GitHub): this {
-		const releasePlease = new ReleasePlease(this).addPlugin({type: 'node-workspace'})
-		this.applyGithubJobNxEnv(
-			releasePlease.releaseWorkflow.workflow,
-			'release-please'
-		)
-		const build = gh.tryFindWorkflow('build')!
-		return this.applyGithubBuildFlow(build)
-	}
-
-	applyRecursive(
-		cb: (project: Project, monorepo: this) => void,
-		includeSelf: boolean = false
-	): this {
-		if (includeSelf) cb(this, this)
-		this.sortedSubProjects.forEach((proj) => cb(proj, this))
-		return this
-	}
-
-	applyGithubJobNxEnv(workflow: github.GithubWorkflow, jobId: string): this {
-		const job = workflow.getJob(jobId) as github.workflows.Job
-		workflow.updateJob(jobId, {
-			...job,
-			env: {
-				...job.env,
-				NX_NON_NATIVE_HASHER: 'true',
-				NX_BRANCH: '${{ github.event.number }}',
-				NX_RUN_GROUP: '${{ github.run_id }}',
-				NX_CLOUD_ACCESS_TOKEN: '${{ secrets.NX_CLOUD_ACCESS_TOKEN }}',
-				CI: 'true',
-			},
-		})
-		return this
-	}
-
-	protected applyGithubBuildFlow(workflow: github.GithubWorkflow): this {
-		this.applyGithubJobNxEnv(workflow, 'build')
-		return this
-	}
-
-	protected applyPackage(nodePackage: javascript.NodePackage): this {
-		nodePackage.addField('type', 'module')
-		this.addNxRunManyTask('stub', {
-			skipCache: true,
-			target: 'stub',
-			parallel: 5,
-			noBail: true,
-			ignoreCycles: true,
-			outputStyle: 'stream',
-		})
-		const postInstall =
-			this.tasks.tryFind('post-install') ?? this.tasks.addTask('post-install')
-		postInstall.exec(
-			NodePackageUtils.command.projen(this.package.packageManager, 'stub')
-		)
-		this.addScripts({
-			postinstall: NodePackageUtils.command.projen(
-				this.package.packageManager,
-				'post-install'
-			),
-		})
-		return this
-	}
-
-	protected applyNpmConfig(npmConfig: javascript.NpmConfig): this {
-		// default '*' to highest resolution.
-		npmConfig.addConfig('resolution-mode', 'highest')
-		return this
-	}
-
-	protected buildEsmBundledTsConfig() {
-		const base = this.buildExtendableTypeScriptConfig('tsconfig.base.json', {
-			skipLibCheck: true,
-			strict: true,
-			strictNullChecks: true,
-			resolveJsonModule: true,
-			noImplicitThis: true,
-			noUnusedLocals: true,
-			noUnusedParameters: true,
-			noFallthroughCasesInSwitch: true,
-			forceConsistentCasingInFileNames: true,
-		})
-		base.file.addOverride('compilerOptions.useDefineForClassFields', true)
-		const esm = this.buildExtendableTypeScriptConfig('tsconfig.esm.json', {
-			module: 'ESNext',
-			target: 'ES2022',
-			lib: ['ES2022'],
-			allowSyntheticDefaultImports: true,
-			esModuleInterop: true,
-		})
-		const bundler = this.buildExtendableTypeScriptConfig(
-			'tsconfig.bundler.json',
-			{
-				moduleResolution: TypeScriptModuleResolution.BUNDLER,
-				allowImportingTsExtensions: true,
-				allowArbitraryExtensions: true,
-				resolveJsonModule: true,
-				isolatedModules: true,
-				verbatimModuleSyntax: true,
-				noEmit: true,
-				jsx: javascript.TypeScriptJsxMode.PRESERVE,
-			}
-		)
-		return javascript.TypescriptConfigExtends.fromTypescriptConfigs([
-			base,
-			esm,
-			bundler,
-		])
-	}
-
-	/**
-	 * Build tsconfig primed for extending.
-	 * @param fileName File name of tsconfig.
-	 * @param options Compiler options.
-	 */
-	buildExtendableTypeScriptConfig(
-		fileName: string,
-		options: javascript.TypeScriptCompilerOptions
-	): javascript.TypescriptConfig {
-		const config = new javascript.TypescriptConfig(this, {
-			fileName,
-			compilerOptions: options,
-		})
-		config.file.addDeletionOverride('include')
-		config.file.addDeletionOverride('exclude')
-		return config
-	}
-
-	addWorkspaceDeps(...dependency: (javascript.NodeProject | string)[]) {
-		return PnpmWorkspace.of(this)!.addWorkspaceDeps(...dependency)
-	}
-
-	preSynthesize() {
-		super.preSynthesize()
-		this.applyCleanTask()
-	}
-}
 
 export class ProjenProject extends cdk.JsiiProject {
 	public readonly projectName: ProjectName
@@ -319,7 +86,7 @@ export class TypescriptProject extends typescript.TypeScriptProject {
 	public readonly tsconfigDev: javascript.TypescriptConfig
 
 	constructor(options: TypeScriptProjectOptions) {
-		const { name, workspaceDeps, tsconfigBase, ...rest } = options
+		const { name, workspaceDeps, tsconfigBase, tsconfig, ...rest } = options
 		const projectName = new ProjectName(name)
 		super({
 			...projectDefaults,
@@ -329,6 +96,7 @@ export class TypescriptProject extends typescript.TypeScriptProject {
 			release: true,
 			authorName: 'arroyoDev-LLC',
 			prettier: true,
+			tsconfig,
 			projenCommand: NodePackageUtils.command.exec(
 				javascript.NodePackageManager.PNPM,
 				'projen'
@@ -352,6 +120,7 @@ export class TypescriptProject extends typescript.TypeScriptProject {
 			...tsconfigPaths,
 			compilerOptions: {
 				outDir: 'dist',
+				...(tsconfig?.compilerOptions ?? {}),
 			},
 			extends: tsconfigBase,
 		})
